removed screenshot1.png (duplicate of kwave-main.png), fixed x86_64 warning in SwapFile.cpp
svn-id: https://kwave.svn.sourceforge.net/svnroot/kwave/trunk@2244 bbf32e5a-961d-0410-9e10-91f32741c31e

+++ libkwave/SwapFile.cpp
@@ -123,7 +123,7 @@ SwapFile *SwapFile::resize(size_t size)
     // resize the file
     //  qDebug("SwapFile::resize(%u)", size);
     m_file.seek(size-1);
-    if (m_file.pos() == size-1) {
+    if (static_cast<qint64>(m_file.pos()) == static_cast<qint64>(size - 1)) {
 	if (size > m_size) {
 	    // growing: mark the new "last byte"
 	    m_file.putChar(0);
